requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
et-xmlfile==1.1.0
Flask==2.0.3
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
numpy==1.22.2
openai==0.14.0
openpyxl==3.0.9
pandas==1.4.1
pandas-stubs==1.2.0.48
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
tqdm==4.62.3
twilio==7.6.0
urllib3==1.26.8
Werkzeug==2.0.3
